requirements.txt
# Core dependencies
telethon==1.28.5
google-generativeai==0.3.1
python-dotenv==1.0.0

# GUI dependencies (for desktop app)
customtkinter==5.2.1
pillow==10.0.0

# Web dependencies
flask==2.0.1
flask-login==0.5.0
flask-pymongo==2.3.0
pymongo[srv]==4.5.0
dnspython==2.4.2
authlib==1.2.1
werkzeug==2.0.1
gunicorn==20.1.0
# bson==0.5.10  # Removed - conflicts with pymongo's bson module
itsdangerous==2.0.1
jinja2==3.0.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.2
requests==2.31.0

# Async support
asyncio==3.4.3

# Deployment
waitress==2.1.2